추상 클래스와 인터페이스 트레이드 오프
