+++ src/main/java/objects/ch02_movie/AmountDiscountPolicy.java
@@ -1,6 +1,6 @@
 package objects.ch02_movie;
 
-public class AmountDiscountPolicy extends DiscountPolicy {
+public class AmountDiscountPolicy extends DefaultDiscountPolicy {
     private Money discountAmount;
 
     public AmountDiscountPolicy(Money discountAmount, DiscountCondition... conditions) {
