+++ src/main/java/objects/ch02_movie/NoneDiscountPolicy.java
@@ -1,8 +1,8 @@
 package objects.ch02_movie;
 
-public class NoneDiscountPolicy extends DiscountPolicy {
+public class NoneDiscountPolicy implements DiscountPolicy {
     @Override
-    protected Money getDiscountAmount(Screening screening) {
+    public Money calculateDiscountAmount(Screening screening) {
         return Money.ZERO;
     }
 }
