+++ src/main/java/objects/ch02_movie/PercentDiscountPolicy.java
@@ -1,6 +1,6 @@
 package objects.ch02_movie;
 
-public class PercentDiscountPolicy extends DiscountPolicy {
+public class PercentDiscountPolicy extends DefaultDiscountPolicy {
     private double percent;
 
     public PercentDiscountPolicy(double percent, DiscountCondition... conditions) {
